pyproject.toml
[tool.poetry]
name = "digital-twin-factory"
version = "0.1.0"
description = "AI-Driven Digital Twin Factory System with Vision Transformer"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
# Core
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"
click = "^8.1.0"

# Vision & ML
torch = "^2.0.0"
torchvision = "^0.15.0"
timm = "^0.9.0"
opencv-python = "^4.8.0"
albumentations = "^1.3.0"
Pillow = "^10.1.0"
scikit-learn = "^1.3.0"
numpy = "^1.24.0"
onnx = "^1.15.0"
onnxruntime = "^1.16.0"
onnx-simplifier = "^0.4.0"

# API & Web
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
requests = "^2.31.0"

# Digital Twin & Simulation
simpy = "^4.1.0"

# Predictive Maintenance
xgboost = "^2.0.0"
pandas = "^2.1.0"
scipy = "^1.11.0"

# Scheduling & Optimization
ortools = "^9.8.0"

# Reinforcement Learning
ray = {extras = ["rllib"], version = "^2.9.0"}
gymnasium = "^0.29.0"
tensorboard = "^2.15.0"

# Utilities
tqdm = "^4.66.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.10.0"
ruff = "^0.1.0"
mypy = "^1.6.0"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
